Change int64 to int32

+++ internal/app/mvm/server.go
@@ -20,8 +20,8 @@ type IMVMService interface {
 	GetProfile(id string) (*model.User, error)
 	GetProfiles(ids []string) ([]*model.User, error)
 	SearchForUsers(searchInput string) ([]*model.User, error)
-	UpsertAvatarSettings(id string, settings map[int64]int64) error
-	GetAvatarSettings(id string) (map[int64]int64, error)
+	UpsertAvatarSettings(id string, settings map[int32]int32) error
+	GetAvatarSettings(id string) (map[int32]int32, error)
 
 	CreateRoom(room *mvmPb.Room) (*mvmPb.Room, error)
 	GetRooms(searchQuery string) ([]*mvmPb.Room, error)
